Cambiando el tipo de state de int a tinyint en la tabla commodity

+++ almacen.sql
@@ -179,10 +179,10 @@ CREATE TABLE store_category(
 
 /*MERCANCIA*/
 CREATE TABLE commodity(
-    commodity_id            INT		        NOT NULL AUTO_INCREMENT PRIMARY KEY,		/*mercancia_id*/
-    brand_id                INT             NOT NULL,									/*marca_id*/
-    commodity_name          VARCHAR(200)    NOT NULL,									/*nombre_mercancia*/
-    state					INT				NOT NULL,									/*estado*/
+    commodity_id            INT		            NOT NULL AUTO_INCREMENT PRIMARY KEY,		/*mercancia_id*/
+    brand_id                INT                 NOT NULL,									/*marca_id*/
+    commodity_name          VARCHAR(200)        NOT NULL,									/*nombre_mercancia*/
+    state					TINYINT				NOT NULL,									/*estado*/
 
     FOREIGN KEY (brand_id)          REFERENCES      brand(brand_id)
 );
